+++ src/accounts/urls.py
@@ -1,11 +1,11 @@
 from django.urls import path
 
-from .views import KakaoLoginView, KakaoLoginCallBackView, LogoutView
-
+from .views.kakao_callback import KakaoLoginCallBackView
+from .views.kakao_login import KakaoLoginView
+from .views.logout import LogoutView
 
 urlpatterns = [
-    path('/auth/logout', LogoutView.as_view()),
-
     path('/auth/kakao', KakaoLoginView.as_view(), name='auth-kakao'),
-    path('/auth/kakao/callback', KakaoLoginCallBackView.as_view(), name='auth-kakao-callback')
+    path('/auth/kakao/callback', KakaoLoginCallBackView.as_view(), name='auth-kakao-callback'),
+    path('/auth/logout', LogoutView.as_view()),
 ]
