+++ src/accounts/views/kakao_login.py
@@ -1,0 +1,20 @@
+from rest_framework import status
+from rest_framework.permissions import AllowAny
+from rest_framework.request import Request
+from rest_framework.response import Response
+from rest_framework.views import APIView
+
+from accounts.services.kakao_login import KakaoLoginService
+from core.authentication import CsrfExemptSessionAuthentication
+
+
+class KakaoLoginView(APIView):
+    authentication_classes = [CsrfExemptSessionAuthentication]
+    permission_classes = [AllowAny]
+
+    def get(self, request: Request) -> Response:
+        headers = {
+            'Location': KakaoLoginService().get_login_url(),
+        }
+        return Response(headers=headers, status=status.HTTP_302_FOUND)
+
